+++ src/actions/install.js
