+++ src/index.js
@@ -1,4 +1,4 @@
-import * as actions from './action'
+import * as actions from './actions'
 import plugins from './plugins'
 
 export default {
