+++ src/plugins/index.js
@@ -1,13 +1,19 @@
 import path from './path'
 import userSettings from './user-settings'
-import webpack from './webpack'
-import karma from './karma'
+import webpackConfig from './webpack-config'
+import karmaConfig from './karma-config'
 
-export default function env (env = {}) {
+/**
+ * Passes through all the available plugins setting up the environment
+ *
+ * @param  {Object} env starting environment
+ * @return {Object} resulting environment constructed by the plugins
+ */
+export default function runPlugins (env = {}) {
   return [
     path,
     userSettings,
-    webpack,
-    karma
+    webpackConfig,
+    karmaConfig
   ].reduce((env, plugin) => plugin(env), env)
 }
