+++ src/plugins/karma-config/index.js
