+++ src/plugins/webpack-config/css-modules-plugin.js
