+++ src/plugins/webpack-config/json-plugin.js
