+++ src/plugins/webpack-config/media-plugin.js
