+++ src/plugins/webpack-config/pages-plugin.js
