+++ src/plugins/webpack-config/scss-plugin.js
